refactor(task): consolidate alerting app into task app (SMP-514)
- Move AlertTask model and enums to task/models.py
- Merge alert serializers and views into task app
- Keep /api/alerting/ URL paths for frontend compatibility
- Fix TaskLinkSerializer ContentType resolution for alerttask
- Register AlertTask in task/admin.py
- Add migration 0005 with SeparateDatabaseAndState
- Remove backend/alerting/ app
