--- backend/alerting/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class AlertingConfig(AppConfig):
-    default_auto_field = "django.db.models.BigAutoField"
-    name = "alerting"
