+++ backend/backend/urls.py
@@ -40,7 +40,7 @@ urlpatterns = [
     path('api/budgets/', include('budget_approval.urls')),  # Add API prefix for frontend compatibility
     path('api/retrospective/', include('retrospective.urls')),
     path('api/core/', include('core.urls')),
-    path('api/alerting/', include('alerting.urls')),
+    path('api/alerting/', include('task.alert_urls')),
     path('api/report/', include('report.urls')),
     path('api/', include('task.urls')),
     path('api/policy/', include('policy.urls')),
