+++ backend/task/serializers.py
@@ -1,8 +1,9 @@
 from rest_framework import serializers
-from task.models import Task, ApprovalRecord, TaskComment, TaskAttachment, TaskHierarchy, TaskRelation
+from task.models import Task, ApprovalRecord, TaskComment, TaskAttachment, TaskHierarchy, TaskRelation, AlertTask
 from core.models import Project, ProjectMember
 from core.utils.project import get_user_active_project
 from django.contrib.auth import get_user_model
+from django.utils import timezone
 from django.contrib.contenttypes.models import ContentType
 from django.db.utils import OperationalError, ProgrammingError
 import logging
@@ -437,7 +438,12 @@ class TaskLinkSerializer(serializers.Serializer):
         data['object_id'] = object_id
 
         try:
-            ct = ContentType.objects.get(model=content_type)
+            # AlertTask was consolidated from alerting app into task app; use app_label
+            # to avoid ContentType ambiguity if stale alerting.alerttask exists.
+            if content_type == 'alerttask':
+                ct = ContentType.objects.get(app_label='task', model='alerttask')
+            else:
+                ct = ContentType.objects.get(model=content_type)
         except ContentType.DoesNotExist:
             raise serializers.ValidationError({
                 'content_type': f"Content type '{content_type}' not found."
@@ -581,3 +587,84 @@ class TaskRelationAddSerializer(serializers.Serializer):
         choices=['causes', 'blocks', 'clones', 'relates_to'],
         required=True
     )
+
+
+class AlertTaskSerializer(serializers.ModelSerializer):
+    """Serializer for AlertTask model."""
+
+    class Meta:
+        model = AlertTask
+        fields = [
+            "id",
+            "task",
+            "alert_type",
+            "severity",
+            "affected_entities",
+            "initial_metrics",
+            "acknowledged_by",
+            "acknowledged_at",
+            "assigned_to",
+            "status",
+            "investigation_notes",
+            "resolution_steps",
+            "related_references",
+            "postmortem_root_cause",
+            "postmortem_prevention",
+            "resolved_at",
+            "created_at",
+            "updated_at",
+        ]
+        read_only_fields = ["id", "created_at", "updated_at"]
+
+    def validate_task(self, task: Task) -> Task:
+        if task.type != "alert":
+            raise serializers.ValidationError(
+                'Alert details can only be created for tasks of type "alert".'
+            )
+        return task
+
+    def _get_task_for_validation(self) -> Task | None:
+        if isinstance(self.instance, AlertTask):
+            return self.instance.task
+        task_id = self.initial_data.get("task")
+        if task_id:
+            try:
+                return Task.objects.get(id=task_id)
+            except Task.DoesNotExist:
+                return None
+        return None
+
+    def validate_assigned_to(self, user: User | None) -> User | None:
+        if user is None:
+            return user
+        task_obj = self._get_task_for_validation()
+        if task_obj is None:
+            return user
+        has_membership = ProjectMember.objects.filter(
+            user=user, project=task_obj.project, is_active=True
+        ).exists()
+        if not has_membership:
+            raise serializers.ValidationError("Assignee must be a member of the task project.")
+        return user
+
+    def validate_acknowledged_by(self, user: User | None) -> User | None:
+        if user is None:
+            return user
+        task_obj = self._get_task_for_validation()
+        if task_obj is None:
+            return user
+        has_membership = ProjectMember.objects.filter(
+            user=user, project=task_obj.project, is_active=True
+        ).exists()
+        if not has_membership:
+            raise serializers.ValidationError(
+                "Acknowledging user must be a member of the task project."
+            )
+        return user
+
+    def validate(self, attrs):
+        status_val = attrs.get("status")
+        resolved_at = attrs.get("resolved_at")
+        if status_val in {"resolved", "closed"} and resolved_at is None:
+            attrs["resolved_at"] = timezone.now()
+        return attrs
