+++ backend/task/tests/test_views.py
@@ -3,7 +3,7 @@ from django.contrib.auth import get_user_model
 from django.urls import reverse
 from rest_framework.test import APIClient
 from rest_framework import status
-from task.models import Task, ApprovalRecord
+from task.models import Task, ApprovalRecord, AlertTask, AlertTask
 from core.models import Project, Organization, Team, AdChannel, ProjectMember
 from budget_approval.models import BudgetPool, BudgetRequest
 from asset.models import Asset
@@ -465,6 +465,28 @@ class TaskAPITest(TestCase):
         self.assertTrue(task.is_linked)
         self.assertEqual(task.task_type, 'retrospectivetask')
         self.assertEqual(task.linked_object, retrospective)
+
+    def test_link_task_to_alert_task(self):
+        """Test linking task to alert task (AlertTask consolidated in task app)"""
+        task = Task.objects.create(
+            summary="Alert Task",
+            type="alert",
+            owner=self.user,
+            project=self.project,
+        )
+        alert_task = AlertTask.objects.create(
+            task=task,
+            alert_type="spend_spike",
+            severity="high",
+        )
+        url = reverse("task-link", kwargs={"pk": task.id})
+        data = {"content_type": "alerttask", "object_id": str(alert_task.id)}
+        response = self.client.post(url, data, format="json")
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        task.refresh_from_db()
+        self.assertTrue(task.is_linked)
+        self.assertEqual(task.task_type, "alerttask")
+        self.assertEqual(task.linked_object, alert_task)
     
     def test_link_task_invalid_content_type(self):
         """Test linking task with invalid content type"""
