+++ backend/task/views.py
@@ -8,8 +8,8 @@ from django.db.models import Q
 from rest_framework.views import APIView
 from rest_framework.parsers import MultiPartParser, FormParser
 from django.shortcuts import get_object_or_404
-from task.models import Task, ApprovalRecord, TaskComment, TaskAttachment, TaskHierarchy, TaskRelation, ApprovalChain
-from task.serializers import TaskSerializer, TaskListSerializer, TaskLinkSerializer, ApprovalRecordSerializer, TaskApprovalSerializer, TaskForwardSerializer, TaskCommentSerializer, TaskAttachmentSerializer, SubtaskAddSerializer, TaskRelationAddSerializer
+from task.models import Task, ApprovalRecord, TaskComment, TaskAttachment, TaskHierarchy, TaskRelation, ApprovalChain, AlertTask
+from task.serializers import TaskSerializer, TaskListSerializer, TaskLinkSerializer, ApprovalRecordSerializer, TaskApprovalSerializer, TaskForwardSerializer, TaskCommentSerializer, TaskAttachmentSerializer, SubtaskAddSerializer, TaskRelationAddSerializer, AlertTaskSerializer
 from django.contrib.auth import get_user_model
 from django.contrib.contenttypes.models import ContentType
 from core.models import ProjectMember, Project
@@ -998,6 +998,89 @@ class TaskAttachmentDownloadView(APIView):
         }
         
         return Response(download_data)
+
+
+class AlertTaskListCreateView(generics.ListCreateAPIView):
+    """
+    GET /api/alerting/alert-tasks/
+    POST /api/alerting/alert-tasks/
+    """
+
+    permission_classes = [IsAuthenticated]
+    serializer_class = AlertTaskSerializer
+
+    def get_queryset(self):
+        user = self.request.user
+        if not user.is_authenticated:
+            return AlertTask.objects.none()
+
+        accessible_project_ids = ProjectMember.objects.filter(
+            user=user,
+            is_active=True,
+        ).values_list("project_id", flat=True)
+
+        queryset = AlertTask.objects.select_related("task", "task__project").filter(
+            task__project_id__in=accessible_project_ids
+        )
+
+        task_id = self.request.query_params.get("task_id")
+        if task_id:
+            queryset = queryset.filter(task_id=task_id)
+
+        status_value = self.request.query_params.get("status")
+        if status_value:
+            queryset = queryset.filter(status=status_value)
+
+        return queryset
+
+    def perform_create(self, serializer):
+        task = serializer.validated_data.get("task")
+        if task is None:
+            raise DRFValidationError({"task": "Task is required for alert details."})
+
+        if task.type != "alert":
+            raise DRFValidationError(
+                {"task": 'Alert details can only be created for tasks of type "alert".'}
+            )
+
+        user = self.request.user
+        has_membership = ProjectMember.objects.filter(
+            user=user,
+            project=task.project,
+            is_active=True,
+        ).exists()
+        if not has_membership:
+            raise PermissionDenied("You do not have access to this task.")
+
+        if hasattr(task, "alert_task"):
+            raise DRFValidationError({"task": "Alert details already exist for this task."})
+
+        serializer.save()
+
+
+class AlertTaskRetrieveUpdateView(generics.RetrieveUpdateAPIView):
+    """
+    GET /api/alerting/alert-tasks/{id}/
+    PATCH /api/alerting/alert-tasks/{id}/
+    """
+
+    permission_classes = [IsAuthenticated]
+    serializer_class = AlertTaskSerializer
+    lookup_field = "id"
+
+    def get_queryset(self):
+        user = self.request.user
+        if not user.is_authenticated:
+            return AlertTask.objects.none()
+
+        accessible_project_ids = ProjectMember.objects.filter(
+            user=user,
+            is_active=True,
+        ).values_list("project_id", flat=True)
+
+        return AlertTask.objects.select_related("task", "task__project").filter(
+            task__project_id__in=accessible_project_ids
+        )
 
 
 @api_view(['GET'])
